Remove overexposed method and move Attribute disk serialise and converter
