+++ concept/manager/ConceptManagerImpl.java
@@ -31,8 +31,8 @@ import grakn.core.concept.impl.RoleImpl;
 import grakn.core.concept.impl.RuleImpl;
 import grakn.core.concept.impl.TypeImpl;
 import grakn.core.concept.structure.ElementFactory;
-import grakn.core.concept.impl.AttributeSerialiser;
-import grakn.core.concept.impl.AttributeValueConverter;
+import grakn.core.core.AttributeSerialiser;
+import grakn.core.core.AttributeValueConverter;
 import grakn.core.core.Schema;
 import grakn.core.kb.concept.api.Attribute;
 import grakn.core.kb.concept.api.AttributeType;
