+++ test-integration/graql/reasoner/atomic/BUILD
@@ -57,8 +57,8 @@ java_test(
         "//dependencies/maven/artifacts/com/google/guava",
         "//dependencies/maven/artifacts/org/hamcrest:hamcrest-library",
         "//graql/reasoner",
-        "//kb/server",
         "//kb/graql/reasoner",
+        "//kb/server",
         "//test-integration/rule:grakn-test-server",
         "//test-integration/util:graql-test-util",
         "@graknlabs_graql//java:graql",
@@ -130,7 +130,6 @@ java_test(
         "//kb/concept/api",
         "//kb/graql/reasoner",
         "//kb/server",
-        "//concept/impl",
         "//test-integration/rule:grakn-test-server",
         "//test-integration/util:graql-test-util",
         "@graknlabs_graql//java:graql",
